requirements.txt

requests
python-dotenv
pydantic
prettytable
pytest
responses
